requirement-delivery-pipeline/requirements.txt
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
networkx>=3.2
pyyaml>=6.0
jinja2>=3.1.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0